fix: drawers

+++ src/modules/hair_order_notes/server/procedures.ts
@@ -41,16 +41,25 @@ export const hairOrderNotesRouter = createTRPCRouter({
 			return c;
 		}),
 	delete: protectedProcedure
-		.input(z.object({ noteId: z.string().cuid2() }))
+		.input(z.object({ id: z.string().cuid2().nullable() }))
 		.mutation(async ({ input }) => {
-			const { noteId } = input;
+			const { id } = input;
 
-			const c = await prisma.hairOrderNote.delete({
+			if (!id) {
+				throw new TRPCError({ code: "NOT_FOUND", message: "Missing id" });
+			}
+
+			const note = await prisma.hairOrderNote.delete({
 				where: {
-					id: noteId,
+					id,
 				},
 			});
-			return c;
+
+			if (!note) {
+				throw new TRPCError({ code: "NOT_FOUND" });
+			}
+
+			return note;
 		}),
 	getNotesByHairOrderId: protectedProcedure
 		.input(z.object({ hairOrderId: z.string().cuid2() }))
